feat(phase-6): move print_iteration_json to cli/output.rs
After phases 0–5, tune still reached back into main.rs for the one
remaining pub(crate) helper: print_iteration_json.  Moving it to
cli/output.rs eliminates that last coupling and formalises the
"keep tune in binary" decision from the phase-6 evaluation.

tune's crate:: imports now all resolve through cli/ modules:
  crate::CliError, crate::LoopArgs, crate::OutputFormat,
  crate::print_iteration_json — none reach into main.rs bodies.
