+++ crates/pid-ctl/src/main.rs
@@ -511,14 +511,6 @@ fn sleep_with_socket(
     }
 }
 
-pub(crate) fn print_iteration_json(record: &pid_ctl::app::IterationRecord) -> Result<(), CliError> {
-    let stdout = io::stdout();
-    let mut handle = stdout.lock();
-    serde_json::to_writer(&mut handle, record)
-        .map_err(|error| CliError::new(3, format!("failed to serialize JSON output: {error}")))?;
-    writeln!(handle).map_err(|error| CliError::new(1, format!("stdout write failed: {error}")))
-}
-
 fn run_status(state_path: &std::path::Path) -> Result<(), CliError> {
     let store = StateStore::new(state_path);
     let snapshot = store
